crew-api/crew-api/requirements.txt
crewai
crewai[tools]
rich
agentops
chromadb 
langchain
langchain_experimental
langchain_openai
langchain_chroma
fastapi
uvicorn
pydantic
requests
python-dotenv